still refactoring

+++ MyGameEngine/MonsterChase/PlayerController.h
@@ -15,25 +15,25 @@ public:
 inline bool PlayerController::moveByOrder(const char order) {
 
 	if (order == 'w' && this->object->pos.y != 0) {
-		int tmp = this->object->pos.y;
+		double tmp = this->object->pos.y;
 		tmp -= 1;
 		this->object->pos.y = tmp;
 		return true;
 	}
 	else if (order == 'a' && this->object->pos.x != 0) {
-		int tmp = this->object->pos.x;
+		double tmp = this->object->pos.x;
 		tmp -= 1;
 		this->object->pos.x = tmp;
 		return true;
 	}
 	else if (order == 's' && this->object->pos.y != 100) {
-		int tmp = this->object->pos.y;
+		double tmp = this->object->pos.y;
 		tmp += 1;
 		this->object->pos.y = tmp;
 		return true;
 	}
 	else if (order == 'd' && this->object->pos.x != 100) {
-		int tmp = this->object->pos.x;
+		double tmp = this->object->pos.x;
 		tmp += 1;
 		this->object->pos.x = tmp;
 		return true;
